+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [v1.11.2] - 2018-2-15
 
 ### Added
+- CUMULUS-1169
+  - Added a `@cumulus/common/StepFunctions` module. It contains functions for querying the AWS
+    StepFunctions API. These functions have the ability to retry when a ThrottlingException occurs.
+  - Added `@cumulus/common/aws.retryOnThrottlingException()`, which will wrap a function in code to
+    retry on ThrottlingExceptions.
+  - Added `@cumulus/common/test-utils.throttleOnce()`, which will cause a function to return a
+    ThrottlingException the first time it is called, then return its normal result after that.
 - CUMULUS-1103 Compare the collection holdings in CMR with Cumulus' internal data store
 - CUMULUS-1099 Add support for UMMG JSON metadata versions > 1.4.
     - If a version is found in the metadata object, that version is used for processing and publishing to CMR otherwise, version 1.4 is assumed.
@@ -36,6 +43,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Cloudformation template overrides now work as expected
 
 ### Changed
+- CUMULUS-1169
+  - Deprecated the `@cumulus/common/step-functions` module.
+  - Updated code that queries the StepFunctions API to use the retry-enabled functions from
+    `@cumulus/common/StepFunctions`
 - CUMULUS-1121
   - Schema validation is now strongly enforced when writing to the database.
     Additional properties are not allowed and will result in a validation error.